data_engine/requirements.txt
h5py>=3.10.0
numpy>=1.24.0
opencv-python>=4.8.0
rosbags>=0.9.16
torch>=2.1.0
torchvision>=0.16.0
matplotlib>=3.8.0
pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0
click>=8.1.0
pytest>=7.4.0
